requirements.txt
asgiref==3.5.2
backports.entry-points-selectable==1.1.0
black==22.6.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cleo==0.8.1
click==8.1.3
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
Django==4.0.6
django-allauth==0.51.0
django-filter==22.1
django-phonenumber-field==6.3.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==3.2
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.2
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==20.9
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.52
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.20
PyJWT==2.4.0
pylev==1.4.0
pyparsing==2.4.7
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.7.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
webencodings==0.5.1
